remove jruby from build matrix since basically our dependencies never work
If we ever get to a state where our dependencies work on jruby but genspec
does not, then we can add jruby back to the matrix. But for some years now
that's proven unlikely. We don't do much that's special in genspec so if
all the dependencies work on jruby I see no reason genspec wouldn't as
well. Famous last words, I know.